+++ pycamel/src/errors/ValidationErrors.py
@@ -1,6 +1,7 @@
 class AbsentValidationItems(Exception):
     """ Raise when nothing has been passed to validator as validation target """
     def __init__(self, *args):
+        """Constructor, forwards all args to the base Exception class."""
         super().__init__(*args)
 
 
@@ -9,6 +10,7 @@ class IncorrectAssertParameter(Exception):
     Raise when incorrect type of filter has been passed to .validate method
     """
     def __init__(self, *args):
+        """Constructor, forwards all args to the base Exception class."""
         super().__init__(*args)
 
 
@@ -17,4 +19,5 @@ class IncorrectValidationPath(Exception):
     Raise when path to validation item is not correct.
     """
     def __init__(self, *args):
+        """Constructor, forwards all args to the base Exception class."""
         super().__init__(*args)